@@ -76,7 +76,6 @@ func RegisterPlugin(name string, factory plugins.Factory) {
 
 // Params stores the configuration for an OPA instance.
 type Params struct {
-
 	// Globally unique identifier for this OPA instance. If an ID is not specified,
 	// the runtime will generate one.
 	ID string
@@ -250,7 +249,6 @@ type Runtime struct {
 // NewRuntime returns a new Runtime object initialized with params. Clients must
 // call StartServer() or StartREPL() to start the runtime in either mode.
 func NewRuntime(ctx context.Context, params Params) (*Runtime, error) {
-
 	if params.ID == "" {
 		var err error
 		params.ID, err = generateInstanceID()
@@ -423,7 +421,6 @@ func (rt *Runtime) StartServer(ctx context.Context) {
 // will block until either: an error occurs, the context is canceled, or
 // a SIGTERM or SIGKILL signal is sent.
 func (rt *Runtime) Serve(ctx context.Context) error {
-
 	if rt.Params.Addrs == nil {
 		return fmt.Errorf("at least one address must be configured in runtime parameters")
 	}
@@ -449,7 +446,6 @@ func (rt *Runtime) Serve(ctx context.Context) error {
 	undo, err := maxprocs.Set(maxprocs.Logger(func(f string, a ...interface{}) {
 		rt.logger.Debug(f, a...)
 	}))
-
 	if err != nil {
 		rt.logger.WithFields(map[string]interface{}{"err": err}).Debug("Failed to set GOMAXPROCS from CPU quota.")
 	}
@@ -610,7 +606,6 @@ func (rt *Runtime) DiagnosticAddrs() []string {
 
 // StartREPL starts the runtime in REPL mode. This function will block the calling goroutine.
 func (rt *Runtime) StartREPL(ctx context.Context) {
-
 	if err := rt.Manager.Start(ctx); err != nil {
 		fmt.Fprintln(rt.Params.Output, "error starting plugins:", err)
 		os.Exit(1)
@@ -633,7 +628,6 @@ func (rt *Runtime) StartREPL(ctx context.Context) {
 		go func() {
 			repl.SetOPAVersionReport(rt.checkOPAUpdate(ctx).Slice())
 		}()
-
 	}
 	repl.Loop(ctx)
 }
@@ -651,7 +645,7 @@ func (rt *Runtime) checkOPAUpdate(ctx context.Context) *report.DataResponse {
 
 func (rt *Runtime) checkOPAUpdateLoop(ctx context.Context, uploadDuration time.Duration, done chan struct{}) {
 	ticker := time.NewTicker(uploadDuration)
-	mr.Seed(time.Now().UnixNano())
+	mr.New(mr.NewSource(time.Now().UnixNano())) // Seed the PRNG.
 
 	for {
 		resp, err := rt.reporter.SendReport(ctx)
@@ -694,7 +688,6 @@ func (rt *Runtime) decisionIDFactory() string {
 }
 
 func (rt *Runtime) decisionLogger(ctx context.Context, event *server.Info) error {
-
 	plugin := logs.Lookup(rt.Manager)
 	if plugin == nil {
 		return nil
@@ -732,7 +725,6 @@ func (rt *Runtime) readWatcher(ctx context.Context, watcher *fsnotify.Watcher, p
 }
 
 func (rt *Runtime) processWatcherUpdate(ctx context.Context, paths []string, removed string) error {
-
 	loaded, err := initload.LoadPaths(paths, rt.Params.Filter, rt.Params.BundleMode, nil, true, false, nil)
 	if err != nil {
 		return err
@@ -741,7 +733,6 @@ func (rt *Runtime) processWatcherUpdate(ctx context.Context, paths []string, rem
 	removed = loader.CleanPath(removed)
 
 	return storage.Txn(ctx, rt.Store, storage.WriteParams, func(txn storage.Transaction) error {
-
 		if !rt.Params.BundleMode {
 			ids, err := rt.Store.ListPolicies(ctx, txn)
 			if err != nil {
@@ -844,7 +835,6 @@ func (rt *Runtime) onReloadLogger(d time.Duration, err error) {
 }
 
 func (rt *Runtime) getWatcher(rootPaths []string) (*fsnotify.Watcher, error) {
-
 	watchPaths, err := getWatchPaths(rootPaths)
 	if err != nil {
 		return nil, err
